[INCOMPATIBLE CHANGE] refactor kustomize-generator options (#135)
options for the kustomize-generator are now passed via a new struct
KustomizeGeneratorOptions, which is supplied to the New*() constructors;
besides providing a suffix for templating-subjected files, it is now
possible to use non-standard template delimiters (i.e. something
different from {{ and }}); in addition, options supplied through the
constructors can be overridden in the source, if there exists a file
.component-config.yaml at the kustomization's root directory;
finally, implicitly generated kustomizations now ignore hidden files,
i.e. files starting with a dot